requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
asttokens=2.2.1=pypi_0
backcall=0.2.0=pypi_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
ccimport=0.4.2=pypi_0
certifi=2022.12.7=py38h06a4308_0
charset-normalizer=2.0.12=pypi_0
chumpy=0.70=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
cumm-cu113=0.4.7=pypi_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
einops=0.6.0=pypi_0
executing=1.2.0=pypi_0
filelock=3.12.0=pypi_0
fire=0.5.0=pypi_0
freetype=2.10.4=h0708190_1
freetype-py=2.4.0=pypi_0
fsspec=2023.5.0=pypi_0
fvcore=0.1.5.post20210915=py38
glfw=2.2.0=pypi_0
google-auth=2.16.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.51.3=pypi_0
huggingface-hub=0.14.1=pypi_0
idna=3.4=pypi_0
imageio=2.9.0=pypi_0
imageio-ffmpeg=0.4.3=pypi_0
imgui=1.3.0=pypi_0
importlib-metadata=6.0.0=pypi_0
iopath=0.1.9=py38
ipython=8.12.2=pypi_0
jedi=0.18.2=pypi_0
jpeg=9e=h166bdaf_1
kiwisolver=1.4.4=pypi_0
lark=1.1.5=pypi_0
lazy-loader=0.1=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.0.10=hc3755c2_1005
lpips=0.1.4=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.4.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mrcfile=1.4.3=pypi_0
munch=3.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.0=pypi_0
ninja=1.10.2=pypi_0
numpy=1.22.4=pypi_0
nvidiacub=1.10.0=0
oauthlib=3.2.2=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=pypi_0
parso=0.8.3=pypi_0
pccm=0.4.6=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py38h06a4308_0
plotly=5.13.1=pypi_0
plyfile=0.7.4=pypi_0
portalocker=2.7.0=pypi_0
pretrainedmodels=0.7.4=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=4.22.0=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pybind11=2.10.3=pypi_0
pyglet=2.0.7=pypi_0
pygments=2.15.1=pypi_0
pyopengl=3.1.0=pypi_0
pyparsing=3.0.9=pypi_0
pyrender=0.1.45=pypi_0
pyspng=0.1.1=pypi_0
python=3.8.16=h7a1cb2a_3
python-dateutil=2.8.2=pypi_0
pytorch-msssim=0.2.1=pypi_0
pytorch3d=0.7.2=dev_0
pywavelets=1.4.1=pypi_0
pyyaml=5.1.2=py38h516909a_1
readline=8.2=h5eee18b_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
scipy=1.9.1=pypi_0
segmentation-models-pytorch=0.3.2=pypi_0
setuptools=65.6.3=py38h06a4308_0
six=1.16.0=pypi_0
smplx=0.1.28=pypi_0
spconv-cu113=2.3.3=pypi_0
sqlite=3.40.1=h5082296_0
stack-data=0.6.2=pypi_0
tabulate=0.9.0=pyhd8ed1ab_1
tenacity=8.2.2=pypi_0
tensorboard=2.12.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.2.0=pyhd8ed1ab_0
tifffile=2023.2.28=pypi_0
timm=0.6.12=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.11.0+cu113=pypi_0
torchaudio=0.11.0=pypi_0
torchvision=0.12.0+cu113=pypi_0
tqdm=4.62.2=pypi_0
traitlets=5.9.0=pypi_0
trimesh=3.21.7=pypi_0
typing-extensions=4.5.0=pypi_0
urllib3=1.26.14=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py38h06a4308_0
xz=5.2.10=h5eee18b_1
yacs=0.1.8=pyhd8ed1ab_0
yaml=0.2.5=h7f98852_2
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=ha95c52a_0
